backend/requirements.txt
sqlalchemy
aiosqlite
fastapi
fastapi-jwt
websocket
python-dotenv
pydantic-settings
authx
alembic
black
asyncpg
psycopg2-binary
uvicorn
